Add compile-based image generation in build_from_source
Adds a generic build script with various build option arguments to build
the accompanying Dockerfile. OS specific Dockerfiles will be placed in
a subdirectory with Debian 12 included in this commit.

Build is done in two stages with the final stage being the "production"
image that will be tagged with the value passed to build.sh script.
Dovecot binaries are located in /dovecot.